check_float_from_listing/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
yarl==1.9.4
